+++ sdk/nexent/core/tools/__init__.py
@@ -1,8 +1,7 @@
-from .bocha_search_tool import BoChaSearchTool
 from .exa_search_tool import EXASearchTool
 from .knowledge_base_search_tool import KnowledgeBaseSearchTool
 from .summary_tool import SummaryTool
 from .send_email_tool import SendEmailTool
 from .get_email_tool import GetEmailTool
 
-__all__ = ["BoChaSearchTool", "EXASearchTool", "KnowledgeBaseSearchTool", "SummaryTool", "SendEmailTool", "GetEmailTool"]
+__all__ = ["EXASearchTool", "KnowledgeBaseSearchTool", "SummaryTool", "SendEmailTool", "GetEmailTool"]
