+++ sdk/nexent/core/tools/knowledge_base_search_tool.py
@@ -17,15 +17,8 @@ class KnowledgeBaseSearchTool(Tool):
                   "Prioritize for company-specific queries. " \
                   "Use for proprietary knowledge or restricted information" \
                   "Avoid for publicly available general knowledge"
-
     inputs = {"query": {"type": "string", "description": "The search query to perform."}}
     output_type = "string"
-
-    messages = {'en': {'search_failed': 'Search request failed: {}',
-        'no_results': 'No results found! Try a less restrictive/shorter query.',
-        'search_success': 'Knowledge Base Search Results'},
-        'zh': {'search_failed': '搜索请求失败：{}', 'no_results': '未找到结果！请尝试使用更宽泛或更短的搜索词。',
-            'search_success': '知识库搜索结果'}}
 
     tool_sign = "a"  # 用于给总结区分不同的索引来源
 
@@ -50,7 +43,6 @@ class KnowledgeBaseSearchTool(Tool):
         self.top_k = top_k
         self.observer = observer
         self.base_url = base_url
-        self.lang = 'zh'
         self.record_ops = 0  # 用于记录序号
 
     def forward(self, query: str) -> str:
@@ -58,13 +50,13 @@ class KnowledgeBaseSearchTool(Tool):
             json={"index_names": self.index_names, "query": query, "top_k": self.top_k})
 
         if kb_search_response.status_code != 200:
-            raise Exception(self.messages[self.lang]['search_failed'].format(kb_search_response.text))
+            raise Exception(f"Search request failed: {kb_search_response.text}")
 
         kb_search_data = kb_search_response.json()
         kb_search_results = kb_search_data["results"]
 
         if not kb_search_results:
-            raise Exception(self.messages[self.lang]['no_results'])
+            raise Exception("No results found! Try a less restrictive/shorter query.")
 
         search_results_json = []  # 将检索结果整理成统一格式
         search_results_return = []  # 输入给大模型的格式
@@ -86,14 +78,3 @@ class KnowledgeBaseSearchTool(Tool):
             search_results_data = json.dumps(search_results_json, ensure_ascii=False)
             self.observer.add_message("", ProcessType.SEARCH_CONTENT, search_results_data)
         return json.dumps(search_results_return, ensure_ascii=False)
-
-
-if __name__ == "__main__":
-    try:
-        tool = KnowledgeBaseSearchTool(index_names=["medical"], base_url="http://localhost:8000", top_k=3)
-
-        question = "乳腺癌的风险"
-        result1 = tool.forward(question)
-        print(result1)
-    except Exception as e:
-        print(e)
