+++ test/sdk/test_exa_search_tool.py
@@ -14,25 +14,13 @@ class TestEXASearchTool(unittest.TestCase):
             exa_api_key=self.api_key,
             observer=self.observer,
             max_results=3,
-            is_model_summary=False,
-            lang='zh'
         )
 
     def test_initialization(self):
         """测试工具初始化"""
         self.assertEqual(self.tool.name, "exa_web_search")
         self.assertEqual(self.tool.max_results, 3)
-        self.assertEqual(self.tool.lang, 'zh')
-        self.assertFalse(self.tool.is_model_summary)
         self.assertFalse(self.tool.image_filter)
-
-    def test_invalid_language(self):
-        """测试不支持的语言设置"""
-        with self.assertRaises(ValueError):
-            EXASearchTool(
-                exa_api_key=self.api_key,
-                lang='invalid_lang'
-            )
 
     @patch('exa_py.Exa.search_and_contents')
     def test_basic_search(self, mock_search):
@@ -68,43 +56,13 @@ class TestEXASearchTool(unittest.TestCase):
         
         self.assertIn("未找到结果", str(context.exception))
 
-    @patch('exa_py.Exa.search_and_contents')
-    def test_model_summary_search(self, mock_search):
-        """测试带模型总结的搜索"""
-        tool_with_summary = EXASearchTool(
-            exa_api_key=self.api_key,
-            is_model_summary=True,
-            lang='zh'
-        )
-
-        mock_result = Mock()
-        mock_result.results = [
-            Mock(
-                title="测试标题",
-                url="http://test.com",
-                text="测试内容",
-                published_date="2024-03-20",
-                extras={"image_links": ["http://test.com/image.jpg"]}
-            )
-        ]
-        mock_search.return_value = mock_result
-
-        result = tool_with_summary.forward("测试查询")
-        
-        # 验证调用时包含summary参数
-        mock_search.assert_called_once()
-        call_kwargs = mock_search.call_args[1]
-        self.assertIn('summary', call_kwargs)
-
     def test_real_search(self):
         """测试真实的搜索调用（不验证结果）"""
         try:
             # 创建一个真实的搜索工具实例
             real_tool = EXASearchTool(
                 exa_api_key=self.api_key,
-                max_results=1,
-                is_model_summary=False,
-                lang='zh',
+                max_results=1
             )
             
             # 执行搜索
